第一阶段完成

+++ src/main/java/com/zewbby/smartticket/domain/entity/TicketOrder.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -30,7 +31,7 @@ public class TicketOrder {
 
     private Integer quantity;
 
-    private Integer totalAmount;
+    private BigDecimal totalAmount;
 
     private String status;
